Move kv classes to separate package
