+++ hedera-mirror-web3/src/test/java/com/hedera/mirror/web3/state/keyvalue/FileReadableKVStateIntegrationTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hedera.mirror.web3.state;
+package com.hedera.mirror.web3.state.keyvalue;
 
 import static org.assertj.core.api.Assertions.assertThat;
